requirements.txt
streamlit==1.31.1
alpaca-trade-api==3.0.2
pandas==2.0.3
plotly==5.18.0
python-dotenv==1.0.0
pytz==2024.1
